mhtcet-advisor/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.26.0
plotly>=5.18.0
reportlab>=4.1.0
pyyaml>=6.0.0
